complete HID Image transition
