+++ DNAnet/data/data_models/hid_image.py
@@ -9,6 +9,10 @@ from typing import Any, Dict, List, MutableMapping, Optional, Sequence, Tuple
 import numpy as np
 from scipy.signal import find_peaks
 
+from DNAnet.data.dataset_compatibility.dataset_strategy import DatasetStrategy, NFI_RND_DatasetStrategy
+from DNAnet.data.kit_compatibility.kit import POWER_PLEX_FUSION_6C_KIT, Kit
+from DNAnet.data.kit_compatibility.scaling_strategy import EPGScalingStrategy, NfiEPGScalingStrategy
+
 from DNAnet.data.data_models import Allele, Annotation, Marker, Panel
 from DNAnet.data.data_models.base import Image
 from DNAnet.data.parsing import get_peak_data, parse_called_alleles
@@ -17,8 +21,8 @@ from DNAnet.data.utils import (
     basepair_interpolator,
     find_peak_boundary,
     find_peak_idx_near_or_in_range,
-    get_interpolated_basepairs,
-    rescale_dye,
+    extract_ss_peaks_new_unify,
+    rescale_dye_new_unify
 )
 from DNAnet.typing import PathLike
 
@@ -34,9 +38,9 @@ class HIDImage(Image):
     :param panel: the panel to be used
     :param annotations_file: the path of the csv/txt file that contains
         the annotations of the HID file.
-    :param include_size_standard: include size standard in data
-        if `true` all six dyes are included
-        if `false` only the first five dyes are included
+    :param include_size_standard: include size standard in the data attribute.
+        if `true` all six dyes are included. For inspection of the HID file.
+        if `false` only the first five dyes are included. For training + testing models.
     :param annotation: any Annotation belonging to the image
     :param use_cache: whether retrieved peaks should be cached
     :param meta: meta information of the HID file.
@@ -45,12 +49,33 @@ class HIDImage(Image):
 
     def __init__(self,
                  path: PathLike,
+                 dataset_strategy: Optional[DatasetStrategy] = None,
+                 scaling_strategy: Optional[EPGScalingStrategy] = None,
+                 kit: Optional[Kit] = None,
                  panel: Optional[Panel] = None,
+                 use_ground_truth_as_annotations: bool = False,
                  annotations_file: PathLike = None,
                  include_size_standard: bool = False,
                  annotation: Optional[Annotation] = None,
                  use_cache: bool = True,
                  meta: MutableMapping[str, Any] = None):
+        
+        # Provide legacy-friendly defaults when strategies are not supplied.
+        if kit is None:
+            kit = POWER_PLEX_FUSION_6C_KIT
+        if panel is None:
+            panel = kit.panel
+
+        if dataset_strategy is None:
+            if panel is None:
+                raise ValueError("Panel must be provided if no dataset_strategy is given.")
+            dataset_strategy = NFI_RND_DatasetStrategy(
+                panel=panel,
+                genotypes_path="resources/data/2p_5p_Dataset_NFI/References",
+            )
+        if scaling_strategy is None:
+            scaling_strategy = NfiEPGScalingStrategy(kit)
+
         self.path = path if isinstance(path, Path) else Path(path)
         self.annotations_file = annotations_file
         self.include_size_standard = include_size_standard
@@ -60,7 +85,11 @@ class HIDImage(Image):
         self._annotation = annotation
         self._meta = meta or dict()
         self._scaler: Optional[np.ndarray] = None
+        self.use_ground_truth_as_annotations = use_ground_truth_as_annotations
         self._panel = panel
+        self.dataset_strategy = dataset_strategy
+        self.scaling_strategy = scaling_strategy
+        self.kit = kit
 
     @property
     def data(self) -> np.ndarray:
@@ -94,21 +123,25 @@ class HIDImage(Image):
             raise FileNotFoundError(str(self.path))
 
         # Parse the raw hid image into a numpy array.
-        if (profile := get_peak_data(self.path)) is None:
+        self.profile = profile = get_peak_data(self.path)
+        if profile is None:
             return None
-        # Use the size standard to translate the location in the profile (array) to base pairs
-        interpolated_base_pairs = get_interpolated_basepairs(np.array(profile[-1]))
-        if interpolated_base_pairs is None:
-            # If the size standard does not pass validation, interpolated_base_pairs
-            # becomes None and the image will be skipped when creating a dataset
+        
+
+        size_standard_dye_lane = np.array(profile[-1])
+        try:
+            ss = self.scaling_strategy.parse_size_standard(size_standard_dye_lane)
+        except ValueError as e:
+            LOGGER.warning(f"Size standard invalid for {self.path.name}: {e}")
             return None
-        # Scale the profile using the size standard
-        data = self._rescale_profile(profile,
-                                     interpolated_base_pairs,
-                                     self.include_size_standard)
-        # Create a scaler, which is used to map a pixel index in the profile to a base pair
-        # location, i.e. the first pixel is in fact BASE_PAIR_START, the last pixel is BASE_PAIR_END
-        self._scaler = interpolated_base_pairs[rescale_dye(interpolated_base_pairs)]
+
+        data = self._rescale_profile(
+            profile,
+            ss.rescaled_indices,
+            self.include_size_standard,
+        )
+        self._scaler = ss.scaler
+        
 
         called_alleles = None
         # Determine the called alleles from the annotations file
@@ -121,8 +154,21 @@ class HIDImage(Image):
         if called_alleles and self.annotation is None:
             # Parse the called alleles into a segmentation
             segmentation = self._get_segmentation(self.scaler, called_alleles, data.shape)
-            self._annotation = Annotation(image=segmentation)
+            self._annotation = Annotation(image=segmentation) # where the annotation is ASSIGNED
             self._meta['called_alleles'] = called_alleles
+
+        # But what if there is no annotations file, only genotype info?
+        # This is ofc hardcoded for the ProvedIt dataset for now
+        if self.use_ground_truth_as_annotations and self.annotation is None and self._panel:
+            try:
+                true_alleles = self.dataset_strategy.load_donor_alleles(self.path.stem)
+                segmentation = self._get_segmentation(self.scaler, true_alleles, data.shape)
+                self._annotation = Annotation(image=segmentation)
+                self._meta["called_alleles"] = true_alleles
+            except ValueError as e:
+                LOGGER.warning(f"Could not load true alleles for {self.path}: {e}")
+                # If we cannot load the true alleles, we do not set the annotation.
+
 
         if data is None:
             raise ValueError(f'Reading {self.path} resulted in None')
@@ -163,23 +209,26 @@ class HIDImage(Image):
             # to avoid missing the scaler when we have not yet read the file.
             self._read()
         return self._scaler[np.newaxis, :]
+    
 
     @staticmethod
-    def _rescale_profile(profile: np.ndarray,
-                         interpolated_base_pairs: np.ndarray,
-                         include_standard: bool) -> np.ndarray:
-        """
-        Rescale profile based on interpolated base pairs.
+    def _rescale_profile(
+        profile: np.ndarray,
+        rescaled_indices: np.ndarray,
+        include_standard: bool,
+    ) -> np.ndarray:
+        """Rescale profile based on precomputed rescale indices.
 
         :param profile: array of dyes in chronological order
-        :param interpolated_base_pairs: the interpolated base pairs
-        :param include_standard: if the size standard should be included
-            in the final profile/data
+        :param rescale_indices: indices of the original profile corresponding to
+            each pixel in the rescaled profile
+        :param include_standard: if the size standard should be included in the
+            final profile/data
         :return: parsed profile as array
         """
         # Select profile based on include_standard flag
         selected_profile = profile if include_standard else profile[:-1]
-        data = selected_profile[:, rescale_dye(interpolated_base_pairs)]
+        data = selected_profile[:, rescaled_indices]
         return data[..., np.newaxis]
 
     @classmethod
@@ -252,7 +301,7 @@ class HIDImage(Image):
 
     def __repr__(self):
         return f"HIDImage({self.path.name})"
-
+    
 
 class Ladder(HIDImage):
     """
