+++ tests/data/data_models/test_provedit_hid_image.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from DNAnet.data.data_models.hid_image import HIDImage
-from DNAnet.data.data_models.provedit_hid_image import ProvedItHIDImage
+from DNAnet.data.data_models.old_hid_image import OldHIDImage as HIDImage
+from DNAnet.data.data_models.hid_image import HIDImage as ProvedItHIDImage
 from DNAnet.data.data_models import Panel
